Question Composition Form
- Addresses the ability to create and edit a question during composition
- Adds hook to check that at least one choice is marked as an answer
- Removed extraneous form from EditQuestion
- Refactored the Choice Bank to be a form; on submit adds a new Choice
Item with the respective input text
- Refactored Choice Items to allow editing after creation
