+++ client/src/utilities/helpers.js
@@ -1,3 +1,6 @@
+import React, { useState } from 'react';
+import shortid from 'shortid';
+
 // Helper method, Compares two arrays for equality
 export function objectsAreEqual(a, b) {
   if (a.length === b.length) {
@@ -7,5 +10,17 @@ export function objectsAreEqual(a, b) {
     }
   }
   return false;
+}
+
+// Renders path elements for an svg
+export function renderSvgPath(paths) {
+  return (
+    paths.map(path => (
+      <path
+        key={shortid.generate()}
+        d={path}
+      />
+    ))
+  );
 }
